cephadm: move context based getters to context_getters.py
Move functions that exist mainly to pull information out of the
CephadmContext in various ways to a new context_getters.py module.
